requirements.txt
# Web Framework
Flask==3.0.0

# HTTP Client (for API calls)
requests==2.31.0

# TLS Fingerprint Bypass (for full job descriptions)
curl_cffi>=0.14.0
beautifulsoup4>=4.9.0
lxml>=4.9.0

# Async Support
asyncio==3.4.3

# Web Scraping (for search results)
playwright==1.48.0

# Database (built-in sqlite3)

# Utilities
python-dotenv==1.0.0

# LLM Integration (NVIDIA NIM)
openai>=1.0.0

# Document Generation
python-docx>=0.8.11
